+++ nitrocli/src/tests/hidden.rs
@@ -31,8 +31,8 @@ fn hidden_create_open_close(model: nitrokey::Model) -> crate::Result<()> {
   {
     let mut manager = nitrokey::force_take()?;
     let device = manager.connect_storage()?;
-    assert!(!device.get_status()?.encrypted_volume.active);
-    assert!(device.get_status()?.hidden_volume.active);
+    assert!(!device.get_storage_status()?.encrypted_volume.active);
+    assert!(device.get_storage_status()?.hidden_volume.active);
   }
 
   let out = ncli.handle(&["hidden", "close"])?;
@@ -41,8 +41,8 @@ fn hidden_create_open_close(model: nitrokey::Model) -> crate::Result<()> {
   {
     let mut manager = nitrokey::force_take()?;
     let device = manager.connect_storage()?;
-    assert!(!device.get_status()?.encrypted_volume.active);
-    assert!(!device.get_status()?.hidden_volume.active);
+    assert!(!device.get_storage_status()?.encrypted_volume.active);
+    assert!(!device.get_storage_status()?.hidden_volume.active);
   }
 
   Ok(())