+++ nitrocli/CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
   - Added `structopt` dependency in version `0.3.7`
   - Replaced `argparse` with `structopt`
   - Removed the `argparse` dependency
+- Bumped `nitrokey` dependency to `0.5.1`
 
 
 0.3.1